Mark registries tests nonparallel
Remove test_registries from original file

Flake8 error
