+++ tests/integration/cattletest/core/test_container.py
@@ -40,93 +40,6 @@ def test_container_build(super_client, context, client):
     assert image.data.fields.build.rm
 
 
-def test_container_registry_default(super_client, client, context,
-                                    admin_user_client):
-
-    id = 'registry.default'
-    default_registry = admin_user_client.by_id_setting(id)
-    default_registry = admin_user_client.update(default_registry,
-                                                value='')
-
-    id = 'registry.whitelist'
-    whitelist_registries = admin_user_client.by_id_setting(id)
-
-    # case 1: whitelist is set to different values, default set to null
-    whitelist_value = 'index.docker.io,sim:rancher,docker.io'
-    whitelist_registries = admin_user_client.update(whitelist_registries,
-                                                    value=whitelist_value)
-    default_registry = admin_user_client.update(default_registry, value='')
-    uuid = "sim:{}".format(random_num())
-    # passes, because no default provided so index.docker.io
-    # gets prepended and that's whitelisted
-    container = client.create_container(imageUuid=uuid,
-                                        startOnCreate=False)
-
-    container = client.wait_success(container)
-
-    assert_fields(container, {
-        "type": "container",
-        "state": "stopped",
-        "imageUuid": uuid,
-        "firstRunning": None,
-    })
-
-    uuid = "sim:localhost.localdomain:5000/ubuntu:{}".format(random_num())
-    # will fail, registry localhost.localdomain:5000 not whitelisted
-    with pytest.raises(ApiError) as e:
-        client.create_container(imageUuid=uuid,
-                                startOnCreate=False)
-    assert e.value.error.status == 422
-    assert e.value.error.code == 'InvalidOption'
-
-    # case 2: whitelist and default both set to different values
-    default_value = 'sim:localhost.localdomain:5000'
-    whitelist_value = 'sim:localhost.localdomain:5000,index.docker.io'
-    whitelist_value = whitelist_value+',sim:rancher,docker.io'
-    default_registry = admin_user_client.update(default_registry,
-                                                value=default_value)
-    whitelist_registries = admin_user_client.update(whitelist_registries,
-                                                    value=whitelist_value)
-    uuid = "sim:v2"
-    # passes, because no default provided so localhost.localdomain:5000
-    # gets prepended and that's whitelisted
-    container = client.create_container(imageUuid=uuid,
-                                        startOnCreate=False)
-
-    container = client.wait_success(container)
-
-    assert_fields(container, {
-        "type": "container",
-        "state": "stopped",
-        "imageUuid": "sim:localhost.localdomain:5000/sim:v2",
-        "firstRunning": None,
-    })
-
-    uuid = "sim:localhost.localdomain:5001/ubuntu:{}".format(random_num())
-    # will fail, registry localhost.localdomain:5001 not whitelisted
-    with pytest.raises(ApiError) as e:
-        client.create_container(imageUuid=uuid,
-                                startOnCreate=False)
-    assert e.value.error.status == 422
-    assert e.value.error.code == 'InvalidOption'
-
-    uuid = "sim:localhost.localdomain:5000/ubuntu:{}".format(random_num())
-    # passes, since localhost.localdomain:5000/ whitelisted
-    container = client.create_container(imageUuid=uuid,
-                                        startOnCreate=False)
-
-    container = client.wait_success(container)
-
-    assert_fields(container, {
-        "type": "container",
-        "state": "stopped",
-        "imageUuid": uuid,
-        "firstRunning": None,
-    })
-    default_registry = admin_user_client.update(default_registry,
-                                                value='')
-
-
 def test_container_create_only(super_client, client, context):
     uuid = "sim:{}".format(random_num())
     container = super_client.create_container(accountId=context.project.id,
